Bug #605 unclear error message with incorrect image url
Authoring: User get unclear error message when tries to create artifact with incorrect image url

Server code changed to return a BadRequest with an error message if a misformed url is entered into an exhibit.  Per talking with Javier the entire exhibit will fail to create if 1 content item fails.