Hotfix - Kiosk Mode Now Defaults to Unchecked
This is a partial fix for Kiosk Mode. Change implemented to default to
unchecked. However, in this version, bug is still present where Kiosk
Mode and other theme values are not populated from the database when the
edit collection form is displayed. This bug is fully resolved in Neil's
MultiUser branch, which will hopefully be deployed within a day or so.
Please note only code changed was one line in
auth-edit-collection-form.ts. Other changes result from
compiling/building.