Display Name of Other Editor
When two or more users have the edit pane open at the same time and have
touched the same exhibit without refreshing to see each others' changes,
everyone after the first person to save is warned that another user has
already changed the exhibit. This change is to improve the warning by
including the user name of the first person to save.