Search: Help Distinguish Between Collections
As content can be copied into different collections, it is probable that
portions of existing curated content can appear in search results
multiple times (if searching over all public collections.) To make it
easier to distinguish which results belongs to which collection when
multiple results have the same title, results belonging to the current
collection are displayed first and all results have hover-over text
detailing who is curating the content and the name of the curator's
collection.

Since we're now exposing the collection name for the first time, the
Beta Content collection was renamed to Cosmos, and the NotDefined user
name (user who owns the default content) was renamed to ChronoZoom. It's
IMPORTANT that the BaseCollectionsAdministrator value (which is also
displayed now) is changed from "NotDefined" to "ChronoZoom" or something
more appropriate in web.config.