Clickable Tour Icons: Production Fix
Found that on production only, clicking on the icon for the tour only
works if an image tile was not generated for the tour and we have the
default chevron icon. If we have an image tile, it is no longer
clickable. This is likely because the underlying default image, which
had the click event, was replaced with a new tile image, and we only
have generated tiles in production. Changed clickable item to be the
image container which should resolve.