Fixed Bibliography Link Color
Bibliography link color (at bottom of round exhibit) was previously the
same color as the timeline border on loading then changed to the exhibit
border after being hovered over. As a result, it sometimes couldn't be
seen if the exhibit background was a similar or the same color as one of
these borders. Fixed by making the bibliography link the same color as
the exhibit title.