UI Updates: New Home Page and My Collections Overlays New Home Page and My Collections overlays to completely replace existing "start page" system per https://trello.com/c/CUGQrodT/147-home-page-my-collections-overlay-general-ui, https://trello.com/c/86b7wBek/189-home-page-overlay-specifics, https://trello.com/c/Gcwk89yj/190-my-collections-overlay-specifics and https://trello.com/c/4xrxKnLq/185-selection-of-new-featured-collections-list.
Final background color and featured item list TBD.

Note: This check-in includes all of the new functionality in a test menu but the old "start page" overlays are still in place. Therefore this version is for testing only, and should not be deployed to production without further changes. (Removal of old "start page" system.)