UI Updates: Remove Start Page, Featured and Tweets Cleanup Changes Made --------------------
1. Removed old start page code and replaced with calls to new overlay code instead.
2. Removed test menu as no longer needed.
3. Removed featured API/code that runs on server-side for featured list.
4. Removed anything to do with Twitter/tweeting.
5. Fixed a few minor bugs encountered during cleanup.

Release Notes
-------------

This code is ready for testing as a potential release, but should *NOT* be pushed to production just yet.
The final Featured Items list is still being debated and we hope to have ideally nine items there.
Additionally, we are trying/debating what color or image to use as the new overlay background still.
So further minor changes will be checked in that will:
- Possibly touch /featured//featured.json
- Possibly include a couple more featured tile images in /featured/
- Probably involve a simple color change in css/overlay.less

In addition, web.config has some changes that should be deployed regarding the new overlays in general:
- Removal of a number of now redundant app keys.
- Addition of one new app key
- Removal of two redundant API references from the services section.