UI Updates: Added Splash Screen
Since the home page overlay backgrounds differ from the Big History
background, it looks kind of messy to partially see the underlying
collection and it's background before rendering the overlay. Plus the
Big History collection can be slow to load on a client so we want to
hide the initial load part. Solution for both was to have a "loading"
splash screen display first until ready to render the overlay or expose
the collection.