Auto-Start Tours Completed
Can now link to and auto-start a tour from a URL by having
@auto-tour={the tour GUID} at the end of the URL, where the part before
the @ consists of a link to the root timeline of the collection that
contains the tour. This includes launching a tour in a different
collection to the one currently displayed. See /featured/featured.json
for an example. (ChronoZoom's introductory tour.)

The Intro Tour link on the home page overlay now remains present, even
when on collections other than the Big History collection.

The ChronoZoom Introductory Tour has been added to the default featured
content list so there are now 7 items. We hope to add two more featured
content items (totalling nine) in the near future.