Collection Export to File Completed
Replaced old file save implementation with library that is more
cross-browser compatible. Tested OK on IE11, latest FireFox and latest
Chrome. Please note did not have a suitably recent Safari to test with.
(Do not have access to a Mac, and will fail on PC version of Safari,
which is a number of years' old.) However, this is webkit based so
should be OK since Chrome is saving correctly.