HOTFIX: YouTube Video Auto-Play
We now notice if autoplay=1 is appended to a YouTube URL, and when
converting the URL to an embedded link, ensure that we append
?autoplay=1 to the link so that the content item will automatically
start to play the video when visited. Previously adding/editing YouTube
video links were converted discarding any autoplay setting, making it
impossible to add new videos that automatically start, such as videos
within tours.