Fix for Re-Opening Customize Collection Pane
Previously when you opened the customize collection pane, it would open
fine the first time but elements would appear to be missing a second
time (without a refresh) and it would freeze on the third attempt.. This
was due to a JS initialization call (which occurs on each open) failing
as the item being initialized had already been initialized the first
time. Added check to see if already initialized which resolves issue.