Updated 2015 Installation Guide
Added additional helpful information gleaned after assisting a class of
students to be set up from scratch for ChronoZoom development.
Improvements to the documentation consist of:

1. Reference to http://jsonlint.com/ as an easier method for prettifying
JSON in the ACS section. (Thanks Fontys!)

2. Reference to ensuring ASP.NET MVC is installed too in the Visual
Studio section. 3 of the 15-20 students had issues since this was not
automatically installed for them.

3. Documentation of why we can not use localhost or port 4949 if needing
to work locally on OneDrive code changes, and how to change your
configuration accordingly.